apps/proxmox-vm-worker/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
deprecation==2.1.0
exceptiongroup==1.2.2
frozenlist==1.5.0
gotrue==2.12.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
multidict==6.2.0
packaging==24.2
pluggy==1.5.0
postgrest==1.0.1
propcache==0.3.1
protobuf==6.30.2
proxmoxer==2.2.0
psycopg2==2.9.10
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.10.1
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
realtime==2.4.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.0
supafunc==0.9.4
temporalio==1.10.0
tomli==2.2.1
types-protobuf==5.29.1.20250403
typing-inspection==0.4.0
typing_extensions==4.13.1
ulid-py==1.1.0
urllib3==2.3.0
websockets==14.2
yarl==1.19.0
